refactor(web): move the actions popover into lib/tools
The popover and its rows drive the tool APIs in lib/tools but lived under
refresh-components/popovers, so the feature was split across two trees.

A `components/` directory rather than one `components.tsx`: the popover
alone is over a thousand lines, and a single file would have been half
again as large.

The barrel is the outside world's entry — only AppInputBar imports any of
this. Inside the directory the files import each other directly, since a
sibling reaching through the barrel is a cycle.
