+++ web/src/lib/tools/components/MCPLineItem.tsx
