+++ web/src/lib/tools/components/index.ts
@@ -1,0 +1,14 @@
+export { default as ActionsPopover } from "@/lib/tools/components/ActionsPopover";
+export type { ActionsPopoverProps } from "@/lib/tools/components/ActionsPopover";
+
+export { default as ActionLineItem } from "@/lib/tools/components/ActionLineItem";
+export type { ActionItemProps } from "@/lib/tools/components/ActionLineItem";
+
+export { default as MCPLineItem } from "@/lib/tools/components/MCPLineItem";
+export type { MCPLineItemProps } from "@/lib/tools/components/MCPLineItem";
+
+export { default as SwitchList } from "@/lib/tools/components/SwitchList";
+export type {
+  SwitchListItem,
+  SwitchListProps,
+} from "@/lib/tools/components/SwitchList";
